varint

+++ src/ecs/components/mod.rs
@@ -3,9 +3,9 @@ use specs::{Component, VecStorage};
 
 #[derive(Debug)]
 pub struct Position {
-    pub x: f64,
-    pub y: f64,
-    pub z: f64,
+    pub x: f32,
+    pub y: f32,
+    pub z: f32,
 }
 impl Component for Position {
     type Storage = VecStorage<Self>;
